third_party/WebKit/LayoutTests/fast/css/modify-cached-detached-sheet-2.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>div{}</style>
<style>div{}</style>
<script>
    var sheet1 = document.styleSheets[0];
    var sheet2 = document.styleSheets[1];

    test(() => {
        sheet1.ownerNode.remove();
        assert_equals(sheet1.cssRules.length, 1, "sheet1 should have length 1.");
        assert_equals(sheet2.cssRules.length, 1, "sheet2 should have length 1.");
    }, "Precondition: both sheets should have length 1 after sheet1 is detached.");

    test(() => {
        sheet2.insertRule("span{}", 0);
        assert_equals(sheet2.cssRules.length, 2, "sheet2 should have length 2 after insertRule.");
        assert_equals(sheet1.cssRules.length, 1, "sheet1 should still have length 1.");
    }, "Check cssRules length after insertRule.");
</script>
